refactor(secretseal): extract shared AES-256-GCM sealer, alias in ssoconfig (#496)
Factor the key-agnostic AES-256-GCM sealer out of identity/internal/ssoconfig
into internal/secretseal so the detection outbound-webhook config can seal its
per-destination signing secrets with the same audited implementation rather
than cloning it (find-prior-art / reuse). ssoconfig re-exports Sealer +
NewSealer as aliases so every identity call site is unchanged; the sealer test
suite moves to secretseal (removing the now-duplicated ssoconfig copy).

go build ./..., go vet -tags integration ./server/identity/... , and the
secretseal + ssoconfig + authz suites pass.
